+++ src/app/app-store.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 
-import { AboutStoreService } from '../about/about-store.service';
-import { ContactStoreService } from '../contact/contact-store.service';
-import { HeroesStoreService } from '../heroes/heroes-store.service';
-import { TutorialsStoreService } from '../tutorials/tutorials-store.service';
+import { AboutStoreService } from './about/about-store.service';
+import { ContactStoreService } from './contact/contact-store.service';
+import { HeroesStoreService } from './heroes/heroes-store.service';
+import { TutorialsStoreService } from './tutorials/tutorials-store.service';
 
 @Injectable({
   providedIn: 'root',
